error.c: Fix cast for printf.
2008-09-01  Jerry DeLisle  <jvdelisle@gcc.gnu.org>

	* runtime/error.c: Fix cast for printf.

From-SVN: r139880

+++ libgfortran/runtime/error.c
@@ -199,13 +199,13 @@ show_locus (st_parameter_common *cmp)
       if (filename != NULL)
 	{
 	  st_printf ("At line %d of file %s (unit = %d, file = '%s')\n",
-		   (int) cmp->line, cmp->filename, cmp->unit, filename);
+		   (int) cmp->line, cmp->filename, (int) cmp->unit, filename);
 	  free_mem (filename);
 	}
       else
 	{
 	  st_printf ("At line %d of file %s (unit = %d)\n",
-		   (int) cmp->line, cmp->filename, cmp->unit);
+		   (int) cmp->line, cmp->filename, (int) cmp->unit);
 	}
       return;
     }
